Buy X items and get Y free items discount

+++ JonasHendrickx.Shop.Contracts/Discounts/CalculateInputModel.cs
@@ -4,6 +4,6 @@
     {
         public decimal TotalPrice { get; set; }
         public decimal UnitPrice { get; set; }
-        public decimal Quantity { get; set; }
+        public uint Quantity { get; set; }
     }
 }
